requirements.txt
python-dotenv==1.0.1
sshtunnel==0.4.0
PyMySQL==1.1.1
azure-identity==1.17.1
azure-storage-file-datalake==12.16.0
pyarrow==17.0.0
